Allow users to specify `client` instead of creating it for them
Currently, you have to specify an authUrl and we construct the client.
This is fine, and still supported, but we do have cases where people
want to use the client outside of a react context and it's nicer
to just let them pass in a client to share.

Additionally, this fixes a case where unmounting and remounting the authprovider
can create extra clients.